hoho

+++ sketch_oct30b/TemperatureSensor.cpp
@@ -8,11 +8,11 @@ TemperatureSensor::TemperatureSensor(const uint8_t pinTrigger, const uint8_t rea
 	digitalWrite(triggerPin, HIGH);
 }
 
-const uint8_t TemperatureSensor::readCelsius()
+const int32_t TemperatureSensor::readCelsius()
 {
 	digitalWrite(triggerPin, HIGH);
 	delayMicroseconds(200);
-	int reading = analogRead(readPin);
+	uint16_t reading = analogRead(readPin);
 	digitalWrite(triggerPin, LOW);
 	// converting that reading to voltage, for 3.3v arduino use 3.3
 	double voltage = reading * 5.0; // * 0.0009775171;
